backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
python-jose
motor
python-jose[cryptography]
passlib[bcrypt]
